Refactored decorators for force_async and safe_async
